+++ crates/cardano/src/ashard/rewards.rs
@@ -3,8 +3,8 @@ use dolos_core::{ChainError, EntityKey};
 use tracing::debug;
 
 use crate::{
-    ewrap::AppliedReward, AccountState, AssignRewards, CardanoDelta, CardanoEntity,
-    LeaderRewardLog, MemberRewardLog,
+    ewrap::{AccountId, AppliedReward, BoundaryWork},
+    AccountState, AssignRewards, CardanoDelta, CardanoEntity, LeaderRewardLog, MemberRewardLog,
 };
 
 #[derive(Default)]
@@ -23,11 +23,11 @@ impl BoundaryVisitor {
     }
 }
 
-impl super::BoundaryVisitor for BoundaryVisitor {
+impl crate::ewrap::BoundaryVisitor for BoundaryVisitor {
     fn visit_account(
         &mut self,
-        ctx: &mut super::BoundaryWork,
-        id: &super::AccountId,
+        ctx: &mut BoundaryWork,
+        id: &AccountId,
         account: &AccountState,
     ) -> Result<(), ChainError> {
         let Some(reward) = ctx.rewards.take_for_apply(&account.credential) else {
@@ -111,7 +111,7 @@ impl super::BoundaryVisitor for BoundaryVisitor {
         Ok(())
     }
 
-    fn flush(&mut self, ctx: &mut super::BoundaryWork) -> Result<(), ChainError> {
+    fn flush(&mut self, ctx: &mut BoundaryWork) -> Result<(), ChainError> {
         let mark_protocol = ctx
             .ending_state()
             .pparams
